requiring package 'lmodern' in iitthesis-extra.sty
Currently, we get several font-shape-substitution-related warnings when compiling, e.g.:

"Font shape `OMS/cmr/m/n' in size <10.95> not available"

and PDFLaTeX substitutes it for another shape. These are mostly resolved by using the lmodern package, which offers shapes in arbitrary size.

This is another commit on the path to achieving 0 compilation errors - but we're not there yet.