abbrevs: example text
Text in the example section (under preliminaries) demonstrates how to
use the \gls family of macros. It is removed when the noabbrevs option
is used, as this prevents the glossaries package from being loaded and
thus \gls cannot be used.

The patch also makes sure the front/abbrevs.tex file is loaded when
this section exists, to define the acronyms it depends on.